requirements.txt
pytest
pytest-playwright
pytest-html
pytest-base-url
pytest-metadata
python-dotenv
playwright
pygments
python-slugify
requests
